+++ CHANGELOG.md
@@ -31,6 +31,7 @@
  - add shadow to copy button #491
  - fix empty state messages in photos collection by theme #466
  - send mail to user when subscription to event is declined #377
+ - send mail to organizers & maintainers when subscription to event is confirmed #377
 
 ## 0.1.0 (2017-12-13)
  - production deployment